pid: Split out pid_types.h
mainline inclusion
from mainline-v6.8-rc1
category: performance

Trimming down sched.h dependencies: we dont't want to include more than
the base types.

Cc: Kees Cook <keescook@chromium.org>
Cc: Andy Lutomirski <luto@amacapital.net>
Cc: Will Drewry <wad@chromium.org>
(cherry picked from commit 6d5e9d63683042a8d344cd5d6f9cf23613864a29)
